update users of GET/SETFIELD()
The last two patches updated GETFIELD() and SETFIELD() to no longer
require the user to specify the mask and shift of a field, and to
remove all _LSH defines and rename any _MASK defines.  There are
some places where the masks were used directly, where the caller
needs to have the _MASK suffix removed.  There are also two users
of SETFIELD() where the field name still has the _MASK suffix
because there is an existing macro with the base name.

Change users of SETFIELD() to include the _MASK suffix where needed.

Change direct users of any mask to remove the _MASK suffix.

Signed-off-by: Dan Streetman <ddstreet@ieee.org>